# Performance/28 aggregation speed