--- a/CoreAnalyticView/src/au/gov/asd/tac/constellation/views/analyticview/aggregators/AppendScoreAggregator.java
+++ b/CoreAnalyticView/src/au/gov/asd/tac/constellation/views/analyticview/aggregators/AppendScoreAggregator.java
@@ -17,6 +17,7 @@ package au.gov.asd.tac.constellation.views.analyticview.aggregators;
 
 import au.gov.asd.tac.constellation.views.analyticview.results.AnalyticResult;
 import au.gov.asd.tac.constellation.views.analyticview.results.ScoreResult;
+import au.gov.asd.tac.constellation.views.analyticview.results.ScoreResult.ElementScore;
 import au.gov.asd.tac.constellation.views.analyticview.utilities.AnalyticException;
 import java.util.List;
 import org.apache.commons.collections4.CollectionUtils;
@@ -41,7 +42,7 @@ public class AppendScoreAggregator implements AnalyticAggregator<ScoreResult> {
         aggregateResult.setIgnoreNullResults(results.stream()
                 .anyMatch(result -> result.getIgnoreNullResults()));
 
-        results.forEach(result -> aggregateResult.combine(result));
+        results.forEach(scoreResult -> aggregateResult.combine(scoreResult, ElementScore::combineReplace));
         
         return aggregateResult;
     }