--- a/CoreAnalyticView/src/au/gov/asd/tac/constellation/views/analyticview/aggregators/AppendScoreAggregator.java
+++ b/CoreAnalyticView/src/au/gov/asd/tac/constellation/views/analyticview/aggregators/AppendScoreAggregator.java
@@ -42,7 +42,7 @@ public class AppendScoreAggregator implements AnalyticAggregator<ScoreResult> {
         aggregateResult.setIgnoreNullResults(results.stream()
                 .anyMatch(result -> result.getIgnoreNullResults()));
 
-        results.forEach(scoreResult -> aggregateResult.combine(scoreResult, ElementScore::combineReplace));
+        results.forEach(scoreResult -> aggregateResult.combine(scoreResult));
         
         return aggregateResult;
     }