--- a/CoreAnalyticView/src/au/gov/asd/tac/constellation/views/analyticview/results/AnalyticResult.java
+++ b/CoreAnalyticView/src/au/gov/asd/tac/constellation/views/analyticview/results/AnalyticResult.java
@@ -80,6 +80,10 @@ public abstract class AnalyticResult<D extends AnalyticData> {
         return Collections.unmodifiableList(result.values().stream().collect(Collectors.toList()));
     }
     
+    public final Map<IdentificationData, D> getResult() {
+        return result;
+    }
+    
     public void add(final D resultData) {
         this.result.put(resultData.getIdentificationData(), resultData);
     }