… CoreAnalyticView/src/au/gov/asd/tac/constellation/views/analyticview/results/AnalyticData.java
… CoreAnalyticView/src/au/gov/asd/tac/constellation/views/analyticview/results/EmptyResult.java
… CoreAnalyticView/src/au/gov/asd/tac/constellation/views/analyticview/results/FactResult.java
… CoreAnalyticView/src/au/gov/asd/tac/constellation/views/analyticview/results/GraphResult.java
… CoreAnalyticView/src/au/gov/asd/tac/constellation/views/analyticview/results/IdentificationData.java
… CoreAnalyticView/src/au/gov/asd/tac/constellation/views/analyticview/results/ScoreResult.java