### Description of the Change

<!--
Refactored the way scoreResults are stored to allow for easier combination of results when multiple analytics are run at once. Then made changes to the way the aggregators work to take advantage of this.
This has resulted in improved aggregation speeds.

-->

### Why Should This Be In Core?

<!--

Improvements to the speed of analytic plugins benefit all users.

-->

### Benefits

Speed benefits when using analytic aggregators. Example speed difference:
When running all centrality analytics and using the append aggregator:
1000/1000 Graph: 14,000 ms -> 300 ms
5000/5000 Graph: 525,475ms -> 442 ms

### Possible Drawbacks

Side effect: More results are now shown, previously when aggregating the result would not be shown if any of the analytic produced a null/0 result. Now they show as long as at least one analytic has a non null/0 result.

### Verification Process

<!--

Comparison of results between modified and unmodified version on all aggregators.
Speed tests run using netbeans profiler.

-->

### Applicable Issues

https://github.com/constellation-app/constellation/issues/28